week8/codebase/requirements.txt
# requirements.txt
streamlit
pandas
faiss-cpu
sentence-transformers
openai
transformers
torch
